New: Add route_table_associations to Geo
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
Adding the ability to codify route_table_associations to Geo.

One thing to note here is that since route table associations don't have
the ability to add tags, the Geo ID is the concatenation of the 2
required attributes for a route table association: the subnet ID and the
route table ID. I concatenated them with a `|` so that it was a bit more
human readable (as both the subnet ID and route table ID use `-` in
their ID's).

Also added a barebones `.irbrc` file that simply requires
`./lib/geoengineer`. This allows you to:
```
bundle exec irb
```
And immediately start using the methods and classes defined in Geo.

How has it been tested:
=======================
Added some tests.

@mentions:
==========
@grahamjenson
@lukedemi