New: Add Aws Vpc DHCP Options Set
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
Added the ability to codify DHCP Options set.

By itself, a pretty standard resource.

Also added a couple additional validations:
- validate_at_least_one_present, which checks that at least on of the
specified attributes has been defined. This is useful when there are
number of attributes that could be defined, and we need at least one.
- validate_only_one_present, which is similar to at_least_one_present,
but checks that exactly one of the attributes is defined.

How has it been tested:
=======================
Added specs

@mentions:
==========
@grahamjenson