New: Add AWS VPN (Private) Gateway
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
Added the ability to codify AWS VPN Gateways, also known as Private
Gateways.

The only thing that really needs scrutiny here is whether or not to
require the `vpc_id`.

The Terraform docs have it as optional, but seems kinda silly to create
one without attaching to a VPC. I could be wrong though, and thus
unfairly limiting Geo.

How has it been tested:
=======================
The usual.