New: Add AWS Network ACL Rule
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
Adding the ability to codify single network ACL rules.

How has it been tested:
=======================
The usual, plus an extra test to make sure the ID was being generated
correctly. Turns out that was important, because my first implementation
was all sorts of wrong.