Fix: Account for resources with no tags
Type of change:
===============
- Bug fix

What changed? ... and Why:
==========================
Original syntax:
```
_geo_id: resource[:tags] ? resource[:tags].select { |x| x[:key] == "Name" }.first[:value] : nil
```
While this would work for cases where resource had or did not have tags,
it would fail for the case where it had tags, but not a "Name" tag.

I modified it to this:
```
_geo_id: resource[:tags].find { |tag| tag[:key] == "Name" }&.dig(:value)
```
Which traded one problem for another - if the resource didn't have ANY
tags defined, it would error.

So I solved with this:
```
_geo_id: resource[:tags]&.find { |tag| tag[:key] == "Name" }&.dig(:value)
```
Which now covers all the bases.

How has it been tested:
=======================
Added an additional spec.

@mentions:
==========
@lukedemi

Jira ticket (if appropriate):
=============================

Screenshots (if appropriate):
=============================