New: Add Lambda Permission to Geo
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
This introduces the concept of a Permission for AWS Lambda's.

My guess is that this resource will require a refactor soonish, due to:
- It being a relatively nascent AWS product
- Becuase of this ^, updates to AWS SDK are quite possible
- Listing remote policies is currently only configured for the lastest
version of a function
- Terraform arguments/state do not map nicely to data returned from the
AWS SDK
- No support for pagination yet, and unlike other resources, there is
high potential for a large number of lambda functions

I also fix a rubocop error in `lib/geoengineer/resource.rb`, and a
couple bugs with the AWS Lambda Alias resource and associated specs.

How has it been tested:
=======================
Added a number of tests, fixed another one