Refactor: validate_cidr_block
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
I expected the method GeoEngineer::Resource#validate_cidr_block to
behave like the other validation methods that returned an array of
errors. However the method returned an array no matter if the cidr block
was invalid or not, and the method consumer was supposed to work out if
the returned data actually signified an error or not.

I didn't like how this worked, so I rewrote it to only return an error
message if the cidr block is actually invalid.

How has it been tested:
=======================
All existing tests pass.

@mentions:
==========
@grahamjenson