New: Add new method
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
While I was creating a new template, I realized that for a number of
cases, I need the terraform ID of a resource if it is known, but it
doesn't exist yet, then I need to the terraform reference to that ID.

So I created a method that does that.