New: Add reset capability
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
So, I was working on a template, and I codified a default path forward
that works for most of our environments. A few however, are different,
and so I tried to override those values using a block. The resource in
question uses tags for its `_geo_id`, so I updated the `Name` tag.
After I did this, I assumed that the `_geo_id` would update, but once
it's been called, the value is static, and the lambda used to calculate
it was gone. That meant that the only way to accomodate this was to move
the value into a template parameter, which seemed like a suboptimal
solution.

Instead, I modified the logic so that we cache the Proc used to
calculate the value, and then added a reset function that allows you to
re-calculate any previously evaluated values using the original logic
with the new resource state.

How has it been tested:
=======================
Added a spec.

@mentions:
==========
@grahamjenson