Refactor: #reset
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
Realized that when I tried to use the `#reset` functionality in a real
world project, it didn't reset the remote resource, which meant that
_terraform_id was not refreshed, which meant that the plan still wasn't
up to date.

So I renamed the `reset` and `eager_load` to `reset_attributes` and
`eager_load_attributes`, and added a `reset` function to
`GeoEngineer::Resource`, which calls `reset_attributes` and also resets
the remote variables such that the resource re-calculates the correct
things.

How has it been tested:
=======================
Added some tests

@mentions:
==========
@grahamjenson