Refactor: Update association resources
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
Since most of the association resources are listed/described via their
parent resource (i.e. VPC, route table, subnet, etc...), under the hood,
Terraform expects certain attributes to be present in the state file so
that it can verify that the remote resource actually exists.

So I updated the `#terraform_state` methods on the resources to add
those attributes to the state output.

How has it been tested:
=======================
By including this code in another repo and running `./geo plan`

@mentions:
==========
@grahamjenson