Refactor: Templates
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
I thought it would be best to extend templating functionality to
environments as well as projects. This is my attempt at doing so in the
least painful, most extensible way. This means that we should aim to
avoid case x then... case y then... situations as much as possible.

The base API should be uniform to the point of minimality, and higher
level methods like `from_template` on `projects` and `environments`
should add additional functionality where necessary.

How has it been tested:
=======================
Todo

@mentions:
==========
@grahamjenson