Refactor: Templates
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
This is a refactor of my refactor.

I opted to extend the `Has...` methodology to `Templates` and `Projects`
as well. Right now, those modules are pretty simple, but it keeps the
mental model simple when dealing with these concepts, and also
eliminates some duplicated logic for `find_template` and
`from_template`. I initially tried putting this in `HasResources`, but
found that too confusing to keep straight.

How has it been tested:
=======================
Existing tests pass...

Will add some actual tests if @grahamjenson approves of this design

@mentions:
==========
@grahamjenson