Other: Exclude auto-generated resources
Type of change:
===============
- Clean up

What changed? ... and Why:
==========================
Routes that point to `local` are automatically created when a route
table is created, and cannot be changed.

Routes that point to a VPC endpoint are automatically created when a VPC
endpoint is associated with a route table, and cannot be modified.

NACL Rules that have the number 32767 are automatically created when a
NACL is created, and cannot be modified.

Therefore, they should not be returned by the `#_fetch_remote_resources`
method.

@mentions:
==========
@grahamjenson