New: Update aws_redshift_cluster
Type of change:
===============
- New feature

What changed? ... and Why:
==========================
Copied the username/password logic from `aws_db_instance` to
`aws_redshift_cluster`. This allows us to skip codifying the username
and password for a cluster if remote resource already exists.

What this means is that in practice, cluster(s) will generally be
created manually either via the AWS CLI, or the AWS Web console, and
then codified. This allows the user to sidestep the issue of having to
store sensitive data in the codifying repo.

@mentions:
==========
@lukedemi