Fix: aws_lambda_alias
Type of change:
===============
- Bug fix

What changed? ... and Why:
==========================
Last PR I was lazy. While that was also a bug, it wasn't the only bug.

Turns out the function_name attribute expecs the function_arn. I fixed
that, and added a test to ensure that it actually works as expected now.

How has it been tested:
=======================
Added a test.

@mentions:
==========
@grahamjenson