New: All environments to have tags
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
In some of the repositories where we are using Geo, we only have
environments, and no projects. However, we still want to apply a set of
org/ProjectName tags to all of the resources. Another use case would be
auto-tagging the environment name on every resource.

We replace `merge_project_tags` with `merge_parent_tags` which then
loops through both `@environment` and `@project`.

How has it been tested:
=======================
Modified existing test cases to cover this new behavior.

@mentions:
==========
@grahamjenson @lukedemi