Refactor: Only log if patches loaded
Type of change:
===============
- Refactor

What changed? ... and Why:
==========================
Right now, every time there are no geo patches, we got an extra log
line. Seeing as the default should be no patches, I think it makes sense
to _only_ log when something non-standard is loaded.

Doing this via a begin/rescue isn't ideal, as we'd be rescueing into the
default state, where we want to no-op. Instead, we check if the file
exists, and then require the file if so.

@mentions:
==========
@lukedemi