Parallelize state file generation
A quick performance test revealed that we could parallelize this portion
of the plan process to reduce overall time by a few seconds.

Given a laptop with 8 logical cores, I observed the following timing
differences for the modified section of code:

Optimization          | Execution time
----------------------|------------------
None (current code)   | 8.97s
4 threads             | 2.58s
6 threads             | 1.67s
8 threads             | 1.32s
16 threads            | 1.44s

Experiment revealed that setting the number of threads to the number of
logical cores gave the optimal reduction in execution time.