Increase parallelism of Terraform commands
Terraform exposes a `-parallelism` flag for the `plan` and `apply`
commands, allowing you to specify the degree of parallelism in the run.

We can take advantage of this to speed up these commands. Running on a
laptop with 8 cores, I observed the following improvements when
specifying the following flag for `terraform plan`:

Flag                  | Execution time (average of 3)
----------------------|------------------------------
None (current code)   | 25.5s
`-parallelism=16`     | 18.8s
`-parallelism=24`     | 15.7s
`-parallelism=32`     | 17.1s

Thus it appears that three times the number of logical cores results in
the most substantial reduction in time on average.

With this in mind, change the default parallelism used by these
Terraform commands.