Add helper for streaming output of a command
In order to support prettier diffs in Geo, we need the ability to
capture a portion of the output of `terraform plan` while at the same
time displaying output so the user isn't left with the impression that
nothing is happening.

Add `shell_exec_stream` which is a version of `shell_exec` that takes a
block, allowing you to easily capture and output stdout/stderr based on
some custom logic, e.g.

    shell_exec_stream(cmd) do |stdout_chunk, stderr_chunk|
      ...
    end

I created a separate helper rather than modify `shell_exec` because it's
not clear how this will evolve, so we'll keep them separate for now. I
also didn't want to change all our uses `shell_exec`.