Improve diff output of plan command
The existing `geo plan` output leaves much to be desired because it
relies on the output `terraform plan`, which leaves much to be desired.

In order to get around this, we parse the output of `terraform plan` to
extract the actual changes and perform some transformations to make it
more readable. Big wins here are handling of JSON and mulitline values
to display them with colorized diffs, and omiting attributes whose
values have not changed.

To support this, we introduce the `diffy` and `treetop` gems. Diffy
generates colorized diffs for us, and Treetop allows us to define a
grammar for the `terraform plan` output.

In the event we can't parse the output, we fall back to displaying the
output verbatim. This way we aren't blocked due to a bug in handling
some unexpected plan output.

We had originally considered using `tfjson`[1] to output the binary plan
file, but that would require us to include vendored binaries compiled
for both Linux and macOS. Thus we opted to go with this less-than-ideal
solution, but which allows us to keep the code pure Ruby.

[1] https://github.com/palantir/tfjson