Add TerraformPlan tests
We want a few sanity checks here to ensure when making changes to the
`TerraformPlan` output helpers or grammar definition that we don't break
parsing of common plan outputs.

This required altering the `TerraformPlan#display` method to accept an
optional output parameter which dictates where it should write output
to. This allows us to test that the output is correct by capturing it in
tests into a `StringIO`.

These tests don't test colorization but are still a good sanity check
nonetheless.