[New] Add AWS ALB to supported resources
What changed? ... and Why:
==========================
Added a wrapper for ALB's and added some tests as well.

The ELB chose to enforce the name being present, which precludes the use
of the name prefix. In practice, that was fine, in part because the name
was then used as the terraform ID. With ALB's, the name is only part of
the ID, and the rest contains what looks to be some random value.

This means that the best way to keep track of the resource is via the
tags, however that means that an extra call is required to grab the
tags.

How has it been tested:
=======================
Added the usual resource tests.