[fix] Handle attributes named 'type' that clash with base functions
The `GeoEngineer::Resource` base class defines a function named `type` which
returns the reource type. This however clashes with resources that have an
attribute named `type`. Currently, this happens with 3 objects:

  * `aws_customer_gateway`
  * `aws_route53_record`
  * `aws_vpn_connection`

The problem that happens is an error setting the `type` attribute. It will throw
an `ArgumentError` saying `wrong number of attributes (1 given, 0 expected)`. On
all, the `type` is required, where not setting it would then result in a
validation error.

This simply adds function for each that will allow setting the value in the
attribute hash.