Additional updates to support Route53 zones and records.
This adds additional fixes to Route53 handling, including proper handling for
zone FQDN handling to include the trailing dot in the terraform id, as well as
outputting the terraform state to handle importing zones attached to multiple
VPCs.

Also a number of updates to the record handling, including updating it to work
as remote resource and populate with existing data. It'll properly retrieve all
records across multiple pages, and also has necessary handling for trailing dot
handling.

On record values, Route53 seems a bit inconsistent with its trailing dot
handling. It seems to work with or without, however doesn't massage the data on
its end.