Updates resources for aws_alb resources to be aws_lb
This updates to be in line with Terraform, where the `aws_alb` resources were
renamed to `aws_lb` to support Network Load Balancers, and then the old names
were simply aliases.

This is backwards compatible with existing instrumentation, but named more
generally to support both application and network load balancers.