Add API client support for GitHub
We want to allow GeoEngineer to be used to codify resources in GitHub
organizations on the public GitHub as well as GitHub Enterprise.

In order to do that, this commit removes some AWS-specific logic from
the `Environment` class so that it doesn't report validation errors when
we set up a GitHub-specific environment.