Added default extended attributes for ALB and Target Groups
This adds the default values for some of the extended attributes on ALBs
and ALB target groups. This prevents erronous plan updates, such as
changing idle_timeout from `""` to `"60"`. That is the default value,
but since it wasn't retrieved by Terraform, it didn't realize it was the
already configured value.