Add cloudflare_record resource
This allows you to define a DNS record in Cloudflare.

Due to the implementation of the Cloudflare Terraform provider[1], we
needed to override the `to_terraform_state` method of the Geo resource
so that `terraform refresh` would work correctly. The refresh needed the
presence of `domain` and `zone_id` attributes in the state file in order
to work, instead of just the record ID.

We also included the `name` attribute in the state file, otherwise it
would think the name was empty.

[1] https://github.com/terraform-providers/terraform-provider-cloudflare/blob/35b7cc5a8f56c8b27007b8e48b3f2251ea89a64f/cloudflare/resource_cloudflare_record.go#L339