Support multiple providers for Cloudflare
Due to the design of the Cloudflair gem, all configuration is global,
making it difficult to connect to multiple backend providers.

If we want this to be generally useful, we need to add support for
multiple providers. Work around the issue by using a mutex to prevent
race conditions of multiple provider requests.