Fix aws_network_acl_rule to properly handle remote resources
This updates `aws_network_acl_rule` to be handled as a remote resource.
It previously had `_fetch_remote_resources` defined, however it didn't
utilize `NullObject.maybe(remote_resource)` for it to be triggered.

This issue with that is it can cause failures when planning because it a
resource will make it through to the terraform state that that then
terraform throws a 400 error for because the resource doesn't exist.
This can mainly happen when the ACL doesn't exist yet.