Add ability to nest API Gateway resources.
Currently, geoengineer sets the `parent_id` on all API Gateway resources to the
root resource (that is, '/'). This forces all APIs into a flat structure when
the more natural structure is a hierarchical organization.

This small change makes it so that users can set the `parent_id` to something
other than the root resource, allowing hierarchical organization of API
resources.