Update terraform_id/geo_id handling on aws_rds_cluster_endpoint
This updates the handling for the terraform_id and geo_id in the
`aws_rds_cluster_endpoint` resource. This was being set to the cluster's
identifier, however this wouldn't work when using the resource for custom
endpoints, as it would create the resources, but then on the next run see them
as missing and then fail on the apply.

The resource can only really be used for custom endpoints and not for the built
in default reader/write, from the looks of the Terraform code. On those, AWS
doesn't return a `db_cluster_endpoint_identifier`, but those ones can't really
be altered within Terraform anyway.