Adds support for additional Direct Connect resources.
This adds additional support for link aggregation groups (LAGs) and private
virtual interfaces.

The hosted virtual interfaces are two support interfaces across accounts, where
one account own the Direct Connect hook up, but is linking to a VPC in another
account. The "acceptor" resource is used on the side with the VPC being
connected to.