Update aws_ses_event_destination to output its state
This updates the `aws_ses_event_destination` resource to output all of its state
in a `to_terraform_state` function. The Terraform resource does not implement a
read function, so import/refresh do not retrieve any remote state. As a result,
any Geo run would attempt to destroy/recreate the resource because it would find
it, but have no state on it.

Rubocop was updated because it complained about the function langth and branch
complexity. Unfortunately, mainly because the resource has subresources and
array, and Terraform has fun handling for those in the state file.