[REFACTOR] Better parsing errors
Primarily, this change is to alert the user to _which_ file may have
errored during GPS loading. Before, you'd just get a cryptic stack trace
about YAML or something.

It also refactors the method mechanics, in an attempt to reduce the
AbcSize under 21, but failed. However, since it was done, I kept the
changes.