[NEW] Add dereferencing
This adds the ability to use node references with a GPS file.

Often it is useful to refer some specific attribute of a resource
created by a node, for example a KMS key ARN.

Currently, you can only reference the KMS node, and so other consuming
nodes would have to hardcode the attribute extraction logic.

This works ok, but breaks down when used for things like policies, where
the reference is being substituted into a JSON document. For this, the
user needs to be explicit about which resource and which attribute.

This PR adds that capability, with the syntax:
<project>:<environment>:<configuration>:<node_type>:<node_name>#<resource>.<attribute>

And generates the appropriate terraform reference.

If no resource is specified, it returns the node itself. This can be
useful if you need to access the nodes attributes, or something that
can't be expressed via terraform references.