Complete AwsApiGatewayModel Interface
This commit completes the implementation of `AwsApiGatewayModel`, previously
this class was missing the `_fetch_remote_resources` implementation and this
produced a `NotImplementedError` when attempting to run `geo plan` against a
plan with such a resource.

It includes two new helper methods, `_fetch_remote_rest_api_models` and
`_remote_rest_api_models` to aid in the implementation of
`_fetch_remote_resources`.