Fix: API Gateway Resource Terraform State Issues
Currently, terraform is unable to recognize `aws_api_gateway_model` and
`api_gateway_request_validator` resources because their primary attributes are
incorrect, namely they are using a reference to the rest_api object instead of
the terraform_id of the rest API object. This was causing terraform to attempt
to repeatedly recreate resources that already existed.

This change fixes this issue by using the correct identifier for these primary
attributes.