Adds resource for awss_vpc_peering_connection_accepter
This adds the resource for aws_vpc_peering_connection_accepter, which is needed
in some use cases around cross region peering, where using the non-accepter
resource will create a duplicate resource which enters the failed state once
created.

With this resource, you need to know the connection ID from the requestor side
already, but it is mostly used for Terraform to be able to find and update the
pre-created resource that is pending acceptance.