Add resource for aws_key_pair
This adds a very simple resource for `aws_key_pair`, used to create public keys
for instances in EC2.

The resource doesn't have any tests, since it doesn't leverage any remote
functionality or anything. All there really is on them is the key name and
public key.

This is ignoring changes to `public_key` and `fingerprint`, so any changes to
the key will require you to delete it to apply it. Having either one unignored
would result in it constantly thinking it needed to recreate the resource.

AWS doesn't return the public key, only the fingerprint. We could build the
fingerprint from the public key, and not ignore the fingerprint, but generating
the fingerprint doesn't appear possible with the stdlib in Ruby. It can be done
with `Net::SSH`, but that is adding a dependency for a scarcly used resource.