Adds ability to use _default as a special environment keyword
This adds the ability to use `_default` as a special keyword for the environment
name so the block defined will be used as a template for all other known
environments, except if already defined.

This can be useful when a configuration has to be applied to all
environments. Rather than having to copy and paste an entry for each
environment, it can be defined once. YAML templating can be useful to minimize
how much is copied, but it becomes tedious as the list of environments grows.

Using _default allows the project to shift from being inclusionary to being
exclusionary. If the resources should exist in most environments, but not all,
and the list of ones to not is fairly static, you can exclude the configuration
from that environment just by defining an empty hash for that environment.

For instance, to exlude a node from a `not-here` environment:

```
_default:
  production:
    test_node:
      name: n1
not-here: {}
```

Ordering doesn't matter, so the blank hash can be either before or after the
_default block.