[FIX] Allow hard coding of policy_arn
This modifies the behavior of the aws_iam_role_policy_attachment
resource. Currently, it requires the user to pass in an GeoEngineer
aws_iam_policy resource, and then it figures out the ARN from that
object.

However, there are times when you already know the ARN, and don't want
to pass in a policy object.