Adds validation to aws_security_group for rules to specify source/dest
This adds a validation the aws_security_group which requires any ingress or
egress rules specify a place for the traffic to be coming from/to. Terraform
will allow you to specify a rule which has just port/protocol and doesn't
specify any cidr_block/ipv6_cidr_block/prefix_list_id/security_group. Each of
these fields are option, but you must specify something in one of them in order
for a rule to be valid and able to be applied.

This also updates some other handling to flatten and unique to include
ipv6_cidr_blocks and prefix_list_ids.