Update IAM role policy attach to handle ARN not existing and TF ref
This updates the `aws_iam_role_policy_attachment` handling to better support two
cases.

First, when to not error when the given policy ARN doesn't exist. This makes it
so Geoengineer won't error, but Terraform later will. The primary case is that
you might be creating the policy and the role attachment in the same run. Even
when adding the depends_on, Geoengineer would error out before it expects it to
already exist.

Second, it updates it to skip fetching when the policy ARN is formatted like a
Terraform reference. This also can happen in cases where creating the policy and
the attachment in the same run, or when populating the ARN with a reference
instead of hard setting the value.

This primarily benefits bootstrapping new things where common base resources are
created, like policies, and attaching them to individual items, all in the same
run.