Update aws_internet_gateway to match remote resources on the VPC id
This changes the aws_internet_gateway resources to match the remote resource
based on the vpc_id rather than the Name tag.

This is done because vpc_id is a required attribute, so it is always present,
and a VPC can only have 1 internet gateway attached to it at a time. This will
allow simple matching and better handle things like changing naming patterns in
name tags.

If the gateway is detached, or it is created but not attached in a run, it will
result in provisioning a new one, however there is no harm in that because
there is not specialized configuration for them. If one is removed, then it
would also be removed from all route tables, so there would be no conflict in
provisioning a new one to attach.