[FIX] Allow CloudHsm Resources to plan correctly
Before, the `_geo_id` was not being set properly, so Geo would try to
recreate the HSM on every plan.

This change fixes that, ensuring the `_geo_id` is set properly.

This was tested out by monkey patching the code in aws-resources and
testing with a `./geo plan` against existing resources.