Updates aws_vpc_endpoint to allow tags and require a name tag
This updates the `aws_vpc_endpoint` resource to enable tags and require it to
have a `Name` tag. Previously, it is likely it didn't support tags, but now it
does.

This is step 1 to including the Name tag in the geo_id so we can have multiple
endpoints per service type within a VPC. This is being done in 2 steps so we can
add Name tags to all existing resources before using it in the geo_id, as that
would want it to provision new endpoints for all existing uses.